added stuff

+++ RestND/Data/DiscountServices.cs
@@ -2,12 +2,15 @@ using MySql.Data.MySqlClient;
 using RestND.Data;
 using RestND.MVVM.Model;
 using RestND.MVVM.Model.Orders;
+using RestND.Validations;
 using System;
 using System.Collections.Generic;
+using System.ComponentModel.DataAnnotations;
 
 public class DiscountService() : BaseService<Discount>(DatabaseOperations.Instance)
 {
- 
+    
+
     #region Get All Discounts
 
     public override List <Discount> GetAll()
@@ -20,7 +23,7 @@ public class DiscountService() : BaseService<Discount>(DatabaseOperations.Instan
         {
             discounts.Add(new Discount
             {
-                Discount_ID = Convert.ToInt32(row["Discount_ID"]),
+                Discount_ID = row["Discount_ID"].ToString(),
                 Discount_Name = row["Discount_Name"].ToString(),
                 Discount_Percentage = Convert.ToDouble(row["Discount_Percentage"])
             });
@@ -32,6 +35,8 @@ public class DiscountService() : BaseService<Discount>(DatabaseOperations.Instan
 
     #region Add Discount
     public override bool Add (Discount d){
+
+
         string query = "INSERT INTO discount (Discount_Name, Discount_Percentage) VALUES (@name, @percentage)";
         
         return _db.ExecuteNonQuery(query,
@@ -53,9 +58,12 @@ public class DiscountService() : BaseService<Discount>(DatabaseOperations.Instan
     #endregion
 
     #region Delete Discount
-    public override bool Delete (int discountId){
+    public override bool Delete (string discountId){
+        if (string.IsNullOrEmpty(discountId)) return false;
         string query = "DELETE FROM discount WHERE Discount_ID = @id";
         return _db.ExecuteNonQuery(query, new MySqlParameter("@id", discountId)) > 0;
     }
     #endregion
+
+
 }
